The float conversion's judgement pass: the level and the engine's helpers
The first of the subsystem passes over what the mechanical conversion left
behind: integers now visibly handed to floats, casts with nothing left to
cross, and comments the pass rewrote and made untrue.

util.h says why getExactTime() keeps its double, in the terms that decide
it: the value grows for as long as the session lasts while what is read
off it is the difference of two readings, and a float's step is already
244 us an hour in.

The tree compiles with no errors. The remaining subsystems are still
being worked through, so the integer-to-float gate still has sites open.

+++ Blocks5/src/arrow.cpp
@@ -11,7 +11,7 @@ Arrow::Arrow(Level& level,
 	warpTo(position);
 	flags = OF_MASSIVE | OF_FIXED | OF_ARROWTYPE;
 	this->dir = dir;
-	shownDir = dir;
+	shownDir = static_cast<float>(dir);
 	dirVel = 0.0f;
 	shownAlpha = 0.6f;
 	counter = 0;
@@ -97,7 +97,7 @@ bool Arrow::changeInEditor(int mod)
 {
 	dir++;
 	dir %= 4;
-	shownDir = dir;
+	shownDir = static_cast<float>(dir);
 
 	return true;
 }
